feature: update NotificationEventHandler to send the NotificationData
