+++ app/src/main/java/com/skgtecnologia/sisem/commons/communication/NotificationEventHandler.kt
@@ -1,5 +1,6 @@
 package com.skgtecnologia.sisem.commons.communication
 
+import com.valkiria.uicomponents.bricks.notification.NotificationData
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.SupervisorJob
@@ -7,15 +8,15 @@ import kotlinx.coroutines.launch
 
 internal object NotificationEventHandler {
 
-    fun publishNotificationEvent(data: String) {
+    fun publishNotificationEvent(data: NotificationData) {
         CoroutineScope(Dispatchers.IO).launch {
             EventBus.publish(data)
         }
     }
 
-    fun subscribeNotificationEvent(onEvent: (data: String) -> Unit) {
+    fun subscribeNotificationEvent(onEvent: (data: NotificationData) -> Unit) {
         CoroutineScope(Dispatchers.Main + SupervisorJob()).launch {
-            EventBus.subscribe<String> { data ->
+            EventBus.subscribe<NotificationData> { data ->
                 onEvent(data)
             }
         }
