+++ app/src/main/java/com/skgtecnologia/sisem/ui/preoperational/create/PreOperationalScreen.kt
@@ -21,6 +21,7 @@ import com.valkiria.uicomponents.action.GenericUiAction
 import com.valkiria.uicomponents.action.UiAction
 import com.valkiria.uicomponents.bricks.banner.OnBannerHandler
 import com.valkiria.uicomponents.bricks.loader.OnLoadingHandler
+import com.valkiria.uicomponents.bricks.notification.NotificationData
 import com.valkiria.uicomponents.bricks.notification.OnNotificationHandler
 import kotlinx.coroutines.launch
 import timber.log.Timber
@@ -35,10 +36,9 @@ fun PreOperationalScreen(
     val viewModel = hiltViewModel<PreOperationalViewModel>()
     val uiState = viewModel.uiState
 
-    var hasNotification by remember { mutableStateOf(false) }
-
+    var notificationData by remember { mutableStateOf<NotificationData?>(null) }
     NotificationEventHandler.subscribeNotificationEvent {
-        hasNotification = true
+        notificationData = it
     }
 
     LaunchedEffect(uiState.navigationModel) {
@@ -65,8 +65,8 @@ fun PreOperationalScreen(
 
     PreOperationalScreenRender(viewModel, modifier)
 
-    OnNotificationHandler(hasNotification) {
-        hasNotification = false
+    OnNotificationHandler(notificationData) {
+        notificationData = null
         if (it.isDismiss.not()) {
             // FIXME: Navigate to MapScreen if is type INCIDENT_ASSIGNED
             Timber.d("Navigate to MapScreen")
